Fix uninitialized memory bug
I do not know the cause, only that malloc+memset was not being enough to
initialize memory, but calloc will do. Maybe it has to do with malloc
overcommitting memory it does not have, and calloc is making sure it
exists. I do not really know.

Changes to be committed:
	modified:   src/fast_list.c
	modified:   src/mooshak.c