Refactor fix for uninitialized memory bug
Got the root cause of the bug: the first element of the array aux_count
was not initialized, and had rubbish. That does not completely explain
the fact that only some of the tests were failing. Only initializing
that value is enough to fix the bug. This fix is to use malloc instead
of a slower calloc.

Changes to be committed:
	modified:   src/fast_list.c
	modified:   src/mooshak.c